+++ backend/repository/transaction_repository.py
@@ -27,9 +27,9 @@ class TransactionRepository:
                                                                     'volume': volume,
                                                                     'transaction_timestamp': transaction_timestamp}).execute()
              if transaction_type == 'BUY':
-                UserStatisticsRepository.withdraw_money(float(volume*price))
+                UserStatisticsRepository.withdraw_money(volume*(price))
              else :
-                UserStatisticsRepository.add_money(float(volume*price))
+                UserStatisticsRepository.add_money(volume*float(price))
 
              return add_transaction
          except Error as e:
